Add Main Game Loop (Huge Update)
Finished linear section with transition point in room with big bed. Built girl's AI so that it: looks at player, chases player with set speed, deals a set amount of damage with effects,  disappears when looked at for set duration, and respawns at a random location within a set radius outside of the players vision (most of the time) after being seen or dealing damage. Update also adds 10 collectable mirror shards that spawn after the game transitions. When all shards are collected a subtitle tells you that you won. In total spent 14 hours on this or something.